+++ stm32/Src/main.c
@@ -118,12 +118,13 @@ int main(void)
   APP_PRINTF("Soil Power Sensor Wio-E5 firmware, compiled on %s %s\n", __DATE__, __TIME__);
 
   // configure sensors
-  //SensorsAdd(ADC_measure);
+  SensorsAdd(ADC_measure);
   //SensorsAdd(SDI12_Teros12Measure);
   //SensorsAdd(Phytos31_measure);
+  
+  //BME280Init();
+  //SensorsAdd(BME280Measure); 
 
-  BME280Init();
-  SensorsAdd(BME280Measure); 
 
 
   /* USER CODE END 2 */